witherspoon: Only rebind ucd9000 if we unbound it

+++ meta-openbmc-machines/meta-openpower/meta-ibm/meta-witherspoon/recipes-phosphor/chassis/vcs-pre-post/ucd_disable_vcs.sh
@@ -5,12 +5,18 @@
 # 10/28/16 PGR -
 # 11/10/16 PGR - unbind/bind UCD driver
 
+ucd_retries=5
+ucd=
+
 # unbind ucd driver to permit i2cset
 ucdpath="/sys/bus/i2c/drivers/ucd9000"
 if [ -e $ucdpath ]
 then
-  ucd=`ls $ucdpath | grep 64`
-  echo $ucd > $ucdpath/unbind
+    ucd=`ls -1 $ucdpath | grep 64`
+    if [ -n "$ucd" ]
+    then
+      echo $ucd > $ucdpath/unbind
+    fi
 fi
 
 # Setup UCD module (in standby) to disable VCS rails from power-on:
@@ -26,12 +32,11 @@ i2cset -y 11 0x64 0x02 0x1A i
 i2cset -y 11 0x64 0x00 0x0F i
 i2cset -y 11 0x64 0x02 0x1A i
 
-# re-bind ucd driver
-ucd_retries="5"
-if [ -e $ucdpath ]; then
-  i=0
-  until [ $i -ge $ucd_retries ] || [ -e $ucdpath/$ucd ]; do
-    i=$((i+1))
+# re-bind ucd driver only if we unbound it (i.e. ucd has been set with a value)
+if [ -a -e $ucdpath -n "$ucd" ]; then
+  j=0
+  until [ $j -ge $ucd_retries ] || [ -e $ucdpath/$ucd ]; do
+    j=$((j+1))
     echo $ucd > $ucdpath/bind || ret=$?
   done
   if [ ! -e $ucdpath/$ucd ]; then exit $ret; fi
